# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a LLM Feedback Analysis API server.",
        "title": "LLM Feedback Analysis API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/analyses": {
            "get": {
                "description": "Retrieve all analyses ordered by creation date (newest first) for the history page",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analyses"
                ],
                "summary": "List all analyses",
                "responses": {
                    "200": {
                        "description": "Analyses retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/responses.AnalysisListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing JWT token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/analyses/latest": {
            "get": {
                "description": "Retrieve the most recent completed analysis for the dashboard",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analyses"
                ],
                "summary": "Get latest analysis",
                "responses": {
                    "200": {
                        "description": "Latest analysis retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/responses.AnalysisResponse"
                        }
                    },
                    "204": {
                        "description": "No analysis found"
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing JWT token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/analyses/{id}": {
            "get": {
                "description": "Retrieve a specific analysis with its topics and analyzed feedbacks with their associated topics",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analyses"
                ],
                "summary": "Get analysis by ID",
                "parameters": [
                    {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000",
                        "description": "Analysis ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Analysis retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/responses.AnalysisDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid analysis ID format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing JWT token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Analysis not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
        "/topics": {
            "get": {
                "description": "Retrieve all predefined topics with feedback count and average rating from the latest analysis",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "topics"
                ],
                "summary": "Get topics with statistics",
                "responses": {
                    "200": {
                        "description": "Topics with stats retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/responses.TopicStatsListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing JWT token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/topics/{topic_enum}": {
            "get": {
                "description": "Retrieve detailed information about a specific topic enum with all associated feedbacks",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "topics"
                ],
                "summary": "Get topic details",
                "parameters": [
                    {
                        "type": "string",
                        "example": "product_functionality_features",
                        "description": "Topic enum value",
                        "name": "topic_enum",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Topic details retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/responses.TopicDetailsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid topic enum",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing JWT token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Topic not found in latest analysis",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        }
    },
    "definitions": {
...
        "responses.AnalysisDetailResponse": {
            "description": "Response payload containing detailed analysis with topics and feedback IDs.",
            "type": "object",
            "properties": {
                "analysis": {
                    "$ref": "#/definitions/responses.AnalysisResponse"
                },
                "feedbacks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.FeedbackWithTopicsResponse"
                    }
                },
                "topics": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.TopicAnalysisResponse"
                    }
                }
            }
        },
        "responses.AnalysisListResponse": {
            "description": "Response payload containing a list of analyses.",
            "type": "object",
            "properties": {
                "analyses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.AnalysisResponse"
                    }
                },
                "total": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "responses.AnalysisResponse": {
            "description": "Response payload containing analysis details.",
            "type": "object",
            "properties": {
                "analysis_duration_ms": {
                    "type": "integer",
                    "example": 5000
                },
                "completed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "failure_reason": {
                    "type": "string"
                },
                "feedback_count": {
                    "type": "integer",
                    "example": 100
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "key_insights": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "model": {
                    "type": "string",
                    "example": "gpt-5-mini"
                },
                "new_feedback_count": {
                    "type": "integer"
                },
                "overall_summary": {
                    "type": "string"
                },
                "period_end": {
                    "type": "string",
                    "example": "2024-01-31T23:59:59Z"
                },
                "period_start": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "previous_analysis_id": {
                    "type": "string"
                },
                "sentiment": {
                    "type": "string",
                    "example": "positive"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "tokens": {
                    "type": "integer",
                    "example": 5000
                }
            }
        },
...
        "responses.FeedbackResponse": {
            "description": "Response payload containing feedback details.",
            "type": "object",
            "properties": {
                "comment": {
                    "description": "Feedback comment text",
                    "type": "string",
                    "example": "Great service!"
                },
                "created_at": {
                    "description": "Creation timestamp",
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "deleted_at": {
                    "description": "Deletion timestamp (if deleted)",
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "id": {
                    "description": "Feedback unique identifier",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "rating": {
                    "description": "Rating value from 1 to 5",
                    "type": "integer",
                    "example": 5
                },
                "updated_at": {
                    "description": "Last update timestamp",
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
        "responses.FeedbackWithTopicsResponse": {
            "description": "Response payload containing feedback details with associated topics.",
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string",
                    "example": "Great product!"
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "rating": {
                    "type": "integer",
                    "example": 5
                },
                "topics": {
                    "description": "Topic enum values",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "responses.TopicAnalysisResponse": {
            "description": "Response payload containing topic analysis details.",
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "feedback_count": {
                    "type": "integer",
                    "example": 10
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "sentiment": {
                    "type": "string",
                    "example": "positive"
                },
                "summary": {
                    "type": "string"
                },
                "topic": {
                    "type": "string",
                    "example": "product_functionality_features"
                },
                "topic_name": {
                    "type": "string",
                    "example": "Product Functionality \u0026 Features"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
        "responses.TopicDetailsResponse": {
            "description": "Response payload containing detailed topic information with feedbacks.",
            "type": "object",
            "properties": {
                "average_rating": {
                    "type": "number",
                    "example": 4.5
                },
                "feedback_count": {
                    "type": "integer",
                    "example": 10
                },
                "feedbacks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.FeedbackResponse"
                    }
                },
                "sentiment": {
                    "type": "string",
                    "example": "positive"
                },
                "summary": {
                    "type": "string"
                },
                "topic": {
                    "type": "string",
                    "example": "product_functionality_features"
                },
                "topic_description": {
                    "type": "string"
                },
                "topic_name": {
                    "type": "string",
                    "example": "Product Functionality \u0026 Features"
                }
            }
        },
        "responses.TopicStatsListResponse": {
            "description": "Response payload containing a list of topic statistics.",
            "type": "object",
            "properties": {
                "topics": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.TopicStatsResponse"
                    }
                },
                "total": {
                    "type": "integer",
                    "example": 13
                }
            }
        },
        "responses.TopicStatsResponse": {
            "description": "Response payload containing topic statistics from the latest analysis.",
            "type": "object",
            "properties": {
                "average_rating": {
                    "type": "number",
                    "example": 4.5
                },
                "feedback_count": {
                    "type": "integer",
                    "example": 10
                },
                "topic": {
                    "type": "string",
                    "example": "product_functionality_features"
                },
                "topic_name": {
                    "type": "string",
                    "example": "Product Functionality \u0026 Features"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token for authentication. Use the format: \"Bearer \u003ctoken\u003e\". This follows RFC 6750 (OAuth 2.0 Bearer Token Usage).",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}